Move setup for command endpoints to its own extension class
